Simplified Bolt transaction management ()
* Refactored transaction management.

* Introduced support for injecting connections.

* Fixed an issue in which the database name might be null if a test relies on the system defaults.

* Fixed an issue which causes the database management service builder to discard the selected implementation.

* Reintroduced response type metrics.

* Moved all transaction checks to Connection.

* Corrected the surfacing of database not found and unavailability errors.

* Fixed a race condition which can cause memory tracking errors to be thrown during connection shutdown.

* Adjusted connection shutdown sequence tests to match new behavior.

* Added exception unwrapping to all relevant transaction interactions.

* Corrected transaction behavior for user controlled closure.

* Replaced IllegalStateException with a status bearing exception to trigger Bolt error handling.

* Fixed a hidden merge conflict.

* Fixed a possible NPE when encountering errors while fetching a routing table.

* Unified the error handling within state machine invocations.

* Corrected the statement termination behavior.

* Corrected another small issue within the Fabric tests.

+++ community/bolt/src/main/java/org/neo4j/bolt/protocol/v40/messaging/request/AbstractStreamingMessage.java
@@ -29,9 +29,9 @@ public abstract class AbstractStreamingMessage implements RequestMessage {
 
     private final MapValue meta;
     private final long n;
-    private final int statementId;
+    private final long statementId;
 
-    public AbstractStreamingMessage(MapValue meta, long n, int statementId) {
+    public AbstractStreamingMessage(MapValue meta, long n, long statementId) {
         this.meta = requireNonNull(meta);
         this.n = n;
         this.statementId = statementId;
@@ -41,7 +41,7 @@ public abstract class AbstractStreamingMessage implements RequestMessage {
         return this.n;
     }
 
-    public int statementId() {
+    public long statementId() {
         return statementId;
     }
